Correct modelcounting to be consistent with paper
In the code `wonModels` is the conjunction of the new spec with the negation of the old spec ("#(¬S /\ S', k) / #(S', k)"). It used to be consistent with the one in the paper, but the change was introduced by `rdegiovanni` in commit `2fa815b`, Feb 7th 2020. That commit also then subtracts the resulting ratio from 1.0, but this is still not equivalent to the formula in the paper.

For example, if S=S', then fitness will be 1.0. If S' is twice as weak as S, then fitness will be 0.0. Any larger and we cap out, but erroneously return 1.0. Also, if S' is disjoint from S, as S' gets stronger (fewer models), the ratio approaches 0, and so the fitness approaches 1!